feat: Fix bug in OTP controller to await email sending
The code changes in `otp.controller.js` fix a bug where the email sending was not awaited, potentially causing the response to be sent before the email was actually sent. By adding the `await` keyword before the `sendOTPEmail` function call, we ensure that the email is sent before sending the response.